pyproject.toml
[tool.poetry]
name = "chronos-foundry"
version = "0.1.0"
description = "Production-grade Chronos forecasting system with covariate integration, incremental training, and AWS orchestration"
authors = ["Rupinder Mazara <mazara.rupinder@gmail.com>"]
readme = "README.md"
license = "MIT"
packages = [{include = "chronos_trainer", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
autogluon = ">=1.4.0"
torch = ">=2.0.0"
pandas = "^2.0.0"
numpy = "^1.24.0"
pyarrow = ">=15.0.0"
pyyaml = "^6.0"
pydantic = "^2.0.0"
click = "^8.0.0"
scikit-learn = "^1.3.0"
catboost = "^1.2.0"
xgboost = "^2.0.0"
lightgbm = "^4.0.0"
scipy = "^1.11.0"
boto3 = "^1.28.0"
joblib = "^1.3.0"

[tool.poetry.group.dev.dependencies]
detect-secrets = "^1.4.0"
pre-commit = "^3.0"
pytest = "^7.0"
pytest-cov = "^4.0"
black = "^22.0"
mypy = "^1.0"
isort = "^5.0"
flake8 = "^5.0"

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311', 'py312']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

